OceanWatch3/server/requirements.txt
fastapi
uvicorn
asyncio
pymongo
requests
pydantic
alive-progress
openai
langchain
langchain-openai
python-dotenv
schedule
tenacity
numpy